# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Web Music Player",
      "description": "API realizzate per il progetto di Ingegneria del Software 2022/23",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "language": "it-it",
    "host": "localhost",
    "basePath": "/",
    "tags": [
      {
        "name": "Utenti",
        "description": "Gestione degli account utente"
      },
      {
        "name": "Brani",
        "description": "Operazioni relative ai brani"
      },
      {
        "name": "Ricerca",
        "description": "Ricerca all'interno della piattaforma"
      },
      {
        "name": "Preferiti",
        "description": "Visualizzazione e modifica dei preferiti"
      }
    ],
    "schemes": ["http"],
    "consumes": ["application/json", "application/x-www-form-urlencoded"],
    "produces": ["application/json"],
    "paths": {
        "/api/auth/registrazione": {
            "post": {
                "tags": [
                    "Utenti"
                ],
                "summary": "Aggiungi un utente al sistema",
                "parameters": [
                    {
                        "name": "Nuovo utente",
                        "in": "body",
                        "required": "true",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "$ref": "#/definitions/Utente/properties/email"
                                },
                                "password": {
                                    "$ref": "#/definitions/Utente/properties/password"
                                },
                                "tipoAccount": {
                                    "$ref": "#/definitions/Utente/properties/tipoAccount"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Utente registrato",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "idUtente": {
                                    "$ref": "#/definitions/Utente/properties/id"
                                },
                                "tokenUtente": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Richiesta non valida"
                    },
                    "409": {
                        "description": "Utente già registrato"
                    }
                }
            }
        },
        "/api/auth/accesso": {
            "post": {
                "tags": [
                    "Utenti"
                ],
                "summary": "Effettua l'accesso al sistema",
                "parameters": [
                    {
                        "name": "Utente registrato",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "$ref": "#/definitions/Utente/properties/email"
                                },
                                "password": {
                                    "$ref": "#/definitions/Utente/properties/password"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Accesso eseguito",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "idUtente": {
                                    "$ref": "#/definitions/Utente/properties/id"
                                },
                                "tokenUtente": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accesso negato"
                    },
                    "404": {
                        "description": "Utente non trovato"
                    }
                }
            }
        },
        "/api/eliminaAccount": {
            "delete": {
                "tags": [
                    "Utenti"
                ],
                "summary": "Rimuovi un account dal sistema",
                "parameters": [
                    {
                        "name": "ID utente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "idUtente": {
                                    "$ref": "#/definitions/Utente/properties/id"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Utente rimosso dal sistema"
                    },
                    "400": {
                        "description": "Richiesta non valida"
                    },
                    "404": {
                        "description": "Utente non trovato"
                    }
                }
            }
        },
        "/api/brano/{idBrano}": {
            "get": {
                "tags": [
                    "Brani"
                ],
                "summary": "Ottieni un brano dal suo ID",
                "parameters": [
                    {
                        "name": "idBrano",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Brano/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Brano trovato",
                        "schema": {
                            "$ref":"#/definitions/Brano"
                        }
                    },
                    "400": {
                        "description": "Richiesta non valida"
                    },
                    "404": {
                        "description": "Brano non trovato"
                    }
                }
            }
        },
        "/api/brano": {
            "post": {
                "tags": [
                    "Brani"
                ],
                "summary": "Aggiungi un brano al sistema",
                "parameters": [
                    {
                        "name": "Nuovo brano",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "nomeBrano": {
                                    "$ref": "#/definitions/Brano/properties/nome"
                                },
                                "idUtente": {
                                    "$ref": "#/definitions/Brano/properties/artista"
                                },
                                "durata": {
                                    "type": "integer",
                                    "minumum": 1,
                                    "example": 1
                                },
                                "tags": {
                                    "$ref": "#/definitions/Brano/properties/tags"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Brano aggiunto al sistema"
                    },
                    "400": {
                        "description": "Richiesta non valida"
                    },
                    "409": {
                        "description": "Un brano con lo stesso nome è già presente"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Brani"
                ],
                "summary": "Modifica un brano presente nel sistema",
                "parameters": [{
                    "name": "Brano aggiornato",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "idBrano": {
                                "$ref": "#/definitions/Brano/properties/id"
                            },
                            "nomeBrano": {
                                "$ref": "#/definitions/Brano/properties/nome"
                            },
                            "tags": {
                                "$ref": "#/definitions/Brano/properties/tags"
                            }
                        }
                    }
                }
                ],
                "responses": {
                    "200": {
                        "description": "Modifica effettuata"
                    },
                    "400": {
                        "description": "Richiesta non valida"
                    },
                    "404": {
                        "description": "Brano non trovato"
                    },
                    "409": {
                        "description": "Un brano con lo stesso nome è già presente"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Brani"
                ],
                "summary": "Rimuovi un brano dal sistema",
                "parameters": [
                    {
                        "name": "ID brano",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "idBrano": {
                                    "$ref": "#/definitions/Brano/properties/id"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Brano rimosso dal sistema"
                    },
                    "400": {
                        "description": "Richiesta non valida"
                    },
                    "404": {
                        "description": "Brano non trovato"
                    }
                }
            }
        },
        "/api/ricerca/{testo}": {
            "get": {
                "tags": [
                    "Ricerca"
                ],
                "summary": "Ricerca un brano all'interno del sistema",
                "parameters": [
                    {
                        "name": "testo",
                        "in": "path",
                        "required": "true",
                        "schema": {
                            "type": "string",
                            "example": "Titolo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ricerca eseguita"
                    }
                }
            }
        },
        "/api/preferiti/{idUtente}": {
            "get": {
                "tags": [
                    "Preferiti"
                ],
                "summary": "Ottieni l'elenco dei preferiti di un utente dal suo ID",
                "parameters": [
                    {
                        "name": "idUtente",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Preferiti/properties/utente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista dei prefetiti trovata",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "idBrani": {
                                    "$ref":"#/definitions/Preferiti/properties/listaBrani"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Richiesta non valida"
                    },
                    "404": {
                        "description": "Utente non trovato"
                    }
                }
            }
        },
        "/api/preferiti/modifica": {
            "patch": {
                "tags": [
                    "Preferiti"
                ],
                "summary": "Aggiungi o rimuovi un brano dai preferiti di un utente",
                "parameters": [
                    {
                        "name": "Preferiti",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "idUtente": {
                                    "$ref": "#/definitions/Preferiti/properties/utente"
                                },
                                "idBrano": {
                                    "$ref": "#/definitions/Brano/properties/id"
                                },
                                "azione": {
                                    "type": "string",
                                    "enum": ["aggiunta", "rimozione"],
                                    "example": "aggiunta"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Modifica effettuata",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "idBrani": {
                                    "$ref":"#/definitions/Preferiti/properties/listaBrani"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Richiesta non valida"
                    },
                    "404": {
                        "description": "Utente non trovato"
                    },
                    "409": {
                        "description": "Azione non valida"
                    }
                }
            }
        }
    },
    "definitions": {
        "Utente": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "639b88a1e5e03f310f6f8294"
                },
                "email": {
                    "type": "string",
                    "example": "silvanus.bordignon@studenti.unitn.it"
                },
                "password": {
                    "type": "string",
                    "example": "PasswordValida&"
                },
                "tipoAccount": {
                    "type": "string",
                    "enum": ["standard", "creator"],
                    "example": "standard"
                }
            },
            "required": [
                "id",
                "email",
                "password",
                "tipoAccount"
            ]
        },
        "Brano": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "63a18edd25800766fed3d137"
                },
                "nome": {
                    "type": "string",
                    "example": "Titolo"
                },
                "artista": {
                    "type": "string",
                    "example": "639b88a1e5e03f310f6f8294"
                },
                "durata": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "uniqueItems": "true",
                    "example": ["Pop", "Rock"]
                }
            },
            "required": [
                "id",
                "nome",
                "artista",
                "durata"
            ]
        },
        "Preferiti": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "63a1dcef0f3c5c6958f9623c"
                },
                "utente": {
                    "type": "string",
                    "example": "639b88a1e5e03f310f6f8294"
                },
                "listaBrani": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "example": "63a18edd25800766fed3d137"
                    },
                    "uniqueItems": "true"
                }
            },
            "required": [
                "id",
                "utente"
            ]
        }
    }
}